comments/requirements.txt
Flask==1.1.1
PyYAML==5.3
PyGithub==1.46
